pyproject.toml
[project]
name = "climate-auto"
version = "0.1.0"
description = "Automated weather data collection for TACOCO project"
readme = "README.md"
authors = [
    { name = "eric_zou", email = "ericakccakcc@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
]

[project.scripts]
climate-auto = "climate_auto.main:main"

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
    "ruff>=0.15.6",
]
